requirements.txt
openai
streamlit
pandas
python-dotenv
matplotlib
seaborn